% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bind.to.intervention}
\alias{bind.to.intervention}
\title{Add Rows to Intervention Table}
\usage{
bind.to.intervention(int.table, params, usedouble)
}
\arguments{
\item{int.table}{Dataframe with Day, New.Re (or New.Double.Time), and Days.of.Smoothing
columns.}

\item{params}{ReactiveValues list of parameters.}

\item{usedouble}{Boolean. TRUE if doubling time is used in the app.}
}
\value{
Dataframe with appended row.
}
\description{
Bind rows to the intervention table.
}
