% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cbind.fill}
\alias{cbind.fill}
\title{Column Bind}
\usage{
\method{cbind}{fill}(...)
}
\arguments{
\item{...}{}
}
\value{
Dataframe.
}
\description{
Same as rbind.fill in plyr, but with columns. Stolen from here:
https://stackoverflow.com/questions/7962267/cbind-a-dataframe-with-an-empty-dataframe-cbind-fill
}
