% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create.beta.vec}
\alias{create.beta.vec}
\title{Creates Vector of Beta Values}
\usage{
create.beta.vec(int.table, gamma, usedouble)
}
\arguments{
\item{int.table}{Dataframe with columns: Day, New.Re (or New.Double.Time), and
Days.of.Smoothing}

\item{gamma}{Numeric.}

\item{usedouble}{Boolean. TRUE if doubling time is used as the metric.}
}
\value{
Vector of beta values to use in the SEIR simulation.
}
\description{
This takes as input an intervention dataframe (int.table). The dataframe should have as
columns 'Day' number, new intervention metric (either New.Re or New.Double.Time)
and the number of days that the intervention is smoothed over (Days.of.Smoothing).
}
\details{
The function also takes as input a gamma value a flag for whether doubling time or is
used as the metric (usedouble).

It returns a vector of beta values that correspond with the intervention dataframe.
}
