% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create.cases.df}
\alias{create.cases.df}
\title{Create Cases Dataframe}
\usage{
create.cases.df(df)
}
\arguments{
\item{df}{Dataframe from the SEIR run.}
}
\value{
Dataframe.
}
\description{
Returns a processed dataframe of cases for presentation in the app.
}
