% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create.hosp.df}
\alias{create.hosp.df}
\title{Create Hospitalization Dataframe}
\usage{
create.hosp.df(df)
}
\arguments{
\item{df}{Dataframe from the SEIR run.}
}
\value{
Dataframe.
}
\description{
Returns a processed dataframe of hospitalizations for presentation in the app.
}
