% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create.res.df}
\alias{create.res.df}
\title{Create Hospital Resource Dataframe}
\usage{
create.res.df(df, hosp_cap, icu_cap, vent_cap)
}
\arguments{
\item{df}{Dataframe from the SEIR run.}

\item{hosp_cap}{Numeric, hospital capacity.}

\item{icu_cap}{Numeric, ICU capacity.}

\item{vent_cap}{Numeric, ventilator capacity.}
}
\value{
Dataframe.
}
\description{
Returns a processed dataframe of hospital resources for presentation in the app.
}
