% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{createTransition}
\alias{createTransition}
\title{Generate transition matrix}
\usage{
createTransition(p.g_g, p.g_icu, p.g_d, p.icu_g, p.icu_icu, p.icu_v,
  p.v_icu, p.v_v, p.v_m, p.g_m = 0, p.icu_m = 0)
}
\arguments{
\item{p.g_g}{Numeric. Transition probability from G -> G state.}

\item{p.g_icu}{Numeric. Transition probability from G -> ICU state.}

\item{p.g_d}{Numeric. Transition probability from G -> D state.}

\item{p.icu_g}{Numeric. Transition probability from ICU -> G state.}

\item{p.icu_icu}{Numeric. Transition probability from ICU -> ICU state.}

\item{p.icu_v}{Numeric. Transition probability from ICU -> V state}

\item{p.v_icu}{Numeric. Transition probability from V -> ICU state.}

\item{p.v_v}{Numeric. Transition probability from V -> V state.}

\item{p.v_m}{Numeric. Transition probability from V -> M state.}

\item{p.g_m}{Numeric. Transition probability from G -> M state.}

\item{p.icu_m}{Numeric. Transition probability from ICU -> M state.}
}
\value{
Transition matrix, to be used in the update step of the Markov model.
}
\description{
This function creates the Markov transition matrix using the transition probabilties.
This transition matrix will be used in the Markov model, as diagramed here:
https://drive.google.com/file/d/1_A3Q6C46z4hx02ayqVPhbWuIWFy6Y8ln/view?usp=sharing
}
