% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{default_params}
\alias{default_params}
\title{Returns to Parameter UI elements for the Shiny App}
\usage{
default_params(model)
}
\arguments{
\item{model}{String, either M0, M1, or M2, depending on the model chosen}
}
\value{
the default parameters for the model used
}
\description{
Returns to Parameter UI elements for the Shiny App
}
