% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find.curr.estimates}
\alias{find.curr.estimates}
\title{Find 'Current' Information}
\usage{
find.curr.estimates(model, N, beta.vector, num.days, num.actual, metric,
  start.exp, params)
}
\arguments{
\item{model}{String, model to use (either M0, M1, M2)}

\item{N}{Numeric, number of people in the area.}

\item{beta.vector}{Vector of numerics, should be the same length as num.days.}

\item{num.days}{Numeric. Number of days to simulate.}

\item{num.actual}{Numeric. Number to match on the curve.}

\item{metric}{tring. Metric to use for matching (currently only compatible with 'Cases'
and 'Hospitalizations').}

\item{start.exp}{Starting number of exposures to simulate.}

\item{params}{Parameters to input into the SEIR function.}
}
\value{
List with the day number match as well as counts for susceptible, exposed,
infected and recovered.
}
\description{
This function first runs the simulation on Day 1, with one person exposed.
}
\details{
Based on the metric provided, it finds the day on the curve that corresponds to the
metric. For example, if the user specifies the number of actual hospitalizations is
100 (num.actual = 100, metric = 'Hospitalizations'), then it finds the day on the
curve that corresponds to 100 hospitalizations.

It then returns a list with that day as well as the number susceptible, exposed,
infected, and recovered on that day.
}
