% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getBetaFromDoubling}
\alias{getBetaFromDoubling}
\title{Get Beta from Doubling Time}
\usage{
getBetaFromDoubling(doubling.time, gamma)
}
\arguments{
\item{doubling.time}{Numeric.}

\item{gamma}{Numeric.}
}
\description{
Returns beta from doubling time using formulas here:
https://penn-chime.phl.io/
}
