% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getBetaFromRe}
\alias{getBetaFromRe}
\title{Get Beta from Re}
\usage{
getBetaFromRe(Re, gamma)
}
\arguments{
\item{Re}{Numeric.}

\item{gamma}{Numeric.}
}
\value{
Numeric.
}
\description{
Get Beta from Re
}
