% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{parameters_page}
\alias{parameters_page}
\title{Returns to Parameter UI elements for the Shiny App}
\usage{
parameters_page(model)
}
\arguments{
\item{model}{String, either M0, M1, or M2, depending on the model chosen}
}
\value{
UI elements for parameters in Shiny App
}
\description{
Returns to Parameter UI elements for the Shiny App
}
