% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_df_download}
\alias{process_df_download}
\title{General process dataframe to download}
\usage{
process_df_download(model = "M0", ...)
}
\arguments{
\item{model}{String, either M0, M1, or M2, depending on the model chosen}

\item{...}{arguments to the specific model chosen}
}
\value{
Processed dataframe
}
\description{
General process dataframe to download
}
