% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_df_download_M1}
\alias{process_df_download_M1}
\title{Process dataframe to download for Model 1}
\usage{
process_df_download_M1(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Dataframe which has been processed.
}
\description{
Processes download dataframe with columns with specified names in a
specified order. For model 1, no processing is currently done.
}
