% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_df_download_M2}
\alias{process_df_download_M2}
\title{Process dataframe to download for Model 2}
\usage{
process_df_download_M2(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Dataframe.
}
\description{
Processes download dataframe with columns with specified names in a
specified order.
}
