% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_params_download}
\alias{process_params_download}
\title{General process parameters for download}
\usage{
process_params_download(model = "M0", ...)
}
\arguments{
\item{model}{String, either M0, M1, or M2, depending on the model chosen}

\item{...}{arguments to the specific model chosen}
}
\value{
Parameters List
}
\description{
General process parameters for download
}
