% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_params_download_M0}
\alias{process_params_download_M0}
\title{Process parameters for export for Model 0}
\usage{
process_params_download_M0(params)
}
\arguments{
\item{params}{ReactiveValues list.}
}
\value{
List of processed parameters.
}
\description{
Process parameters for export for Model 0
}
