% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_params_download_M2}
\alias{process_params_download_M2}
\title{Process parameters for export for Model 2}
\usage{
process_params_download_M2(params)
}
\arguments{
\item{params}{ReactiveValues list.}
}
\value{
List of processed parameters.
}
\description{
Process parameters for export for Model 2
}
