% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{roundNonDateCols}
\alias{roundNonDateCols}
\title{Round Non-Date Columns}
\usage{
roundNonDateCols(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
Dataframe.
}
\description{
Helper function in the app that rounds any function that does
not have 'date' as the column name.
}
