% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{runMarkov}
\alias{runMarkov}
\title{Run Markov Simulation}
\usage{
runMarkov(new.g.vec, trans.mat)
}
\arguments{
\item{new.g.vec}{Vector of numerics, which describe the number of new hospitalizations
per day.}

\item{trans.mat}{Matrix. For Markov simulation.}
}
\value{
Dataframe, with counts of each state.
}
\description{
This function takes in as inputs a vector of new hospitalizations per day and
runs the Markov simulation using this.
}
\details{
It returns a dataframe with the number of people in the general hospital (G.state),
non-ventilated ICU (ICU.state), ventilated ICU (V.state), discharged (DCH.state),
and dead (M.state).
}
