% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SEIR}
\alias{SEIR}
\title{Model Caller}
\usage{
SEIR(model = "M0", ...)
}
\arguments{
\item{model}{String, either M0, M1, or M2, depending on the model chosen}

\item{...}{arguments to the specific model chosen}
}
\value{
Dataframe, with susceptible counts, exposed counts, infected counts,
hospitalization numbers, and recovery numbers numbers.
}
\description{
Model Caller
}
