% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SEIR_M1}
\alias{SEIR_M1}
\title{Run SEIR model for Model 1}
\usage{
SEIR_M1(S0, E0, I0, R0, beta.vector, num.days, influx, params)
}
\arguments{
\item{S0}{Numeric. Initial number of susceptible individuals.}

\item{E0}{Numeric. Initial number of exposed individuals.}

\item{I0}{Numeric. Initial number of infected individuals.}

\item{R0}{Numeric. Initial number of recovered individuals.}

\item{beta.vector}{Vector, of beta values to use in the simulation on each day.
Should have the same length as num.days.}

\item{num.days}{Numeric. Number of days to run the simulation.}

\item{influx}{List. This is hash consisting of a 'day' and an 'influx' number,
meant to represent an influx of infections into the region}

\item{params}{List. This should contain the parameters for running the SEIR, including:
- Sigma
- Gamma
- hosp.delay.time: time between infection and hospitalization
- hosp.rate: percent hospitalized out of those infected
- hosp.los: hospital length of stay if not admitted to ICU
- icu.delay.time: time between hospitalization and ICU admission
- icu.rate: percent ICU admitted among those hospitalized
- icu.los : average ICU length of stay if not ventilated
- vent.delay.time: time between ICU ventilation and being put on a ventilator
- vent.rate: percent ventilated among those ICU admitted
- vent.los: average time on a ventilator}
}
\value{
Dataframe, with susceptible counts, exposed counts, infected counts,
hospitalization numbers, and recovery numbers.
}
\description{
Runs the SEIR model using the Model 1 specification.
https://docs.google.com/document/d/1CEoKQ1pD1x4yH7GzR3mViWtfVG-4Do8ViZUd4z_t9Ts/edit?usp=sharing
}
\details{
The simulation ignores births and deaths.
}
