% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SEIR_M2}
\alias{SEIR_M2}
\title{Run SEIR model for Model 2}
\usage{
SEIR_M2(S0, E0, I0, R0, beta.vector, num.days, influx, params)
}
\arguments{
\item{S0}{Numeric. Initial number of susceptible individuals.}

\item{E0}{Numeric. Initial number of exposed individuals.}

\item{I0}{Numeric. Initial number of infected individuals.}

\item{R0}{Numeric. Initial number of recovered individuals.}

\item{beta.vector}{Vector, of beta values to use in the simulation on each day.
Should have the same length as num.days.}

\item{num.days}{Numeric. Number of days to run the simulation.}

\item{influx}{List. This is hash consisting of a 'day' and an 'influx' number,
meant to represent an influx of infections into the region}

\item{params}{List. This should contain the parameters for running the SEIR,
including sigma, gamma.r, gamma.h, hosp.rate, and the transition probabilities
(p.g_g, p.g_icu, p.g_d)}
}
\value{
Dataframe, with susceptible counts, exposed counts, infected counts,
hospitalization numbers, recovery numbers, and death numbers.
}
\description{
Runs the SEIR model, incorporating the separate hospitalization Markov component.
More about the technique is here:
https://drive.google.com/file/d/157yB7O62xyJps2_HUtAelqsQ8e9feTj_/view?usp=sharing
https://sites.me.ucsb.edu/~moehlis/APC514/tutorials/tutorial_seasonal/node4.html
}
\details{
The simulation ignores births and deaths.
}
