% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{data}
\name{start.exp.default}
\alias{start.exp.default}
\title{Server part}
\format{An object of class \code{numeric} of length 1.}
\usage{
start.exp.default
}
\description{
Server part
}
\keyword{datasets}
