% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create.graph}
\alias{create.graph}
\title{Graphing Helper}
\usage{
create.graph(df.to.plot, selected, plot.day, curr.date, frozen_data = NULL)
}
\arguments{
\item{df.to.plot}{Dataframe with data to plot.}

\item{selected}{Vector of strings, representing the columns to plot.}

\item{plot.day}{Date selected to show a vertical line.}

\item{curr.date}{Date where to start the graph.}

\item{frozen_data}{Melted dataframe of previous frozen projections}
}
\value{
ggplot graph.
}
\description{
Graphing helper for the app. Returns a ggplot graph.
}
